Расформирование файлов
Изменено:
- Файл main.py был разбит на несколько файлов.
- Другие небольшие исправления.
